feat: adding heartbeat mechanism for monitoring
Introduces a heartbeat mechanism, which makes a GET request to a
heartbeat URL.

Works using the scheduler mechanism:
On start of the bot, it will check if there is already a heartbeat job
configured. If not, it creates it if the URL and interval is set. If the
job exists, it checks if the interval is still the same as before, if
not, it recreates the job with the new interval.